fix: move logica to authService
