+++ server/src/main/java/com/goofy/controllers/AuthenticationController.java
@@ -1,5 +1,7 @@
 package com.goofy.controllers;
 
+import com.goofy.services.AuthenticationService;
+import com.goofy.services.AuthenticationServiceImpl;
 import com.goofy.services.UserService;
 import com.goofy.services.UserServiceImpl;
 import jakarta.validation.Valid;
@@ -18,18 +20,18 @@ import java.util.Map;
 @RequestMapping("/auth")
 public class AuthenticationController {
 
-    private final UserService userService;
-    private final UserServiceImpl userServiceImpl;
+    private final AuthenticationService authService;
+    private final AuthenticationServiceImpl authServiceImpl;
     @PutMapping("/verify2FA")
     public ResponseEntity<String> verify2FA(@RequestBody @Valid Map<String, String> requestBody, Principal principal) throws Exception {
         String secret = requestBody.get("secret");
         String code = requestBody.get("code");
 
-        return userService.verify2FA(secret, code, principal.getName());
+        return authService.verify2FA(secret, code, principal.getName());
     }
 
     @PutMapping("/disable2FA")
     public ResponseEntity<String> disable2FA(@RequestBody() String uid) throws Exception {
-        return userService.disable2FA(uid);
+        return authService.disable2FA(uid);
     }
 }
