+++ server/src/main/java/com/goofy/services/AuthenticationService.java
@@ -1,4 +1,13 @@
 package com.goofy.services;
 
-public class AuthenticationService {
+import jakarta.validation.Valid;
+import org.springframework.http.ResponseEntity;
+
+import java.util.concurrent.ExecutionException;
+
+public interface AuthenticationService {
+
+    ResponseEntity<String> verify2FA(@Valid String secret, String code, String uid) throws InterruptedException, ExecutionException;
+
+    ResponseEntity<String> disable2FA(String uid) throws InterruptedException, ExecutionException;
 }
