+++ server/src/main/java/com/goofy/services/UserService.java
@@ -19,7 +19,4 @@ public interface UserService {
 
     ResponseEntity<String> changeUsername(@Valid String username, String uid) throws InterruptedException, ExecutionException;
 
-    ResponseEntity<String> verify2FA( @Valid String secret, String code, String uid) throws InterruptedException, ExecutionException;
-
-    ResponseEntity<String> disable2FA(String uid) throws InterruptedException, ExecutionException;
 }
